Split tests into many binaries and move support module to a separate crate
commit-id:6f631d8a
