+++ scarb/tests/build.rs
@@ -5,9 +5,9 @@ use assert_fs::TempDir;
 use indoc::indoc;
 use predicates::prelude::*;
 
-use crate::support::command::Scarb;
-use crate::support::fsx::ChildPathEx;
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::fsx::ChildPathEx;
+use scarb_test_support::project_builder::ProjectBuilder;
 
 #[test]
 fn compile_simple() {
