+++ scarb/tests/build_starknet_contract_allowed_libfuncs.rs
@@ -1,8 +1,8 @@
 use assert_fs::prelude::*;
 use indoc::indoc;
 
-use crate::support::command::Scarb;
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::project_builder::ProjectBuilder;
 
 const EXPERIMENTAL_LIBFUNC: &str = indoc! {r#"
     extern fn redeposit_gas() implicits(GasBuiltin) nopanic;
