+++ scarb/tests/clean.rs
@@ -1,7 +1,7 @@
 use assert_fs::prelude::*;
 
-use crate::support::command::Scarb;
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::project_builder::ProjectBuilder;
 
 #[test]
 fn simple() {
