--- scarb/tests/e2e/main.rs
@@ -1,21 +1,0 @@
-mod add;
-mod build;
-mod build_cairo_plugin;
-mod build_starknet_contract;
-mod build_starknet_contract_allowed_libfuncs;
-mod build_targets;
-mod clean;
-mod flock;
-mod fmt;
-mod git_source;
-mod git_source_network;
-mod manifest_path;
-mod metadata;
-mod new_and_init;
-mod profiles;
-mod remove;
-mod resolver_with_git;
-mod scripts;
-mod subcommand;
-pub(crate) mod support;
-mod test;
