+++ scarb/tests/flock.rs
@@ -8,8 +8,8 @@ use indoc::indoc;
 use io_tee::TeeReader;
 use ntest::timeout;
 
-use crate::support::command::Scarb;
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::project_builder::ProjectBuilder;
 
 #[test]
 #[timeout(60_000)]
