+++ scarb/tests/git_source_network.rs
@@ -5,8 +5,8 @@ use std::thread;
 use assert_fs::TempDir;
 use indoc::{formatdoc, indoc};
 
-use crate::support::command::Scarb;
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::project_builder::ProjectBuilder;
 
 #[test]
 fn https_something_happens() {
