+++ scarb/tests/test_subcommand.rs
@@ -1,9 +1,9 @@
 use assert_fs::TempDir;
 use indoc::{formatdoc, indoc};
 
-use crate::support::command::Scarb;
-use crate::support::filesystem::{path_with_temp_dir, write_script};
-use crate::support::project_builder::ProjectBuilder;
+use scarb_test_support::command::Scarb;
+use scarb_test_support::filesystem::{path_with_temp_dir, write_script};
+use scarb_test_support::project_builder::ProjectBuilder;
 
 #[test]
 #[cfg_attr(
