+++ utils/scarb-test-support/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "scarb-test-support"
+version = "1.0.0"
+edition.workspace = true
+publish = false
+
+[dependencies]
+anyhow.workspace = true
+assert_fs.workspace = true
+camino.workspace = true
+indoc.workspace = true
+itertools.workspace = true
+once_cell.workspace = true
+scarb = { path = "../../scarb" }
+semver.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+snapbox.workspace = true
+toml_edit.workspace = true
+url.workspace = true
