+++ utils/scarb-test-support/src/fsx.rs
@@ -1,12 +1,16 @@
+use std::fs::File;
+use std::io::BufReader;
+
 use assert_fs::fixture::ChildPath;
 use assert_fs::TempDir;
 use camino::Utf8Path;
 use itertools::Itertools;
+use serde::de::DeserializeOwned;
 
 pub use internal_fsx::{PathBufUtf8Ext, PathUtf8Ext};
 
 #[allow(unused)]
-#[path = "../../../src/internal/fsx.rs"]
+#[path = "../../../scarb/src/internal/fsx.rs"]
 mod internal_fsx;
 
 pub trait AssertFsUtf8Ext {
@@ -39,6 +43,7 @@ impl AssertFsUtf8Ext for &ChildPath {
 
 pub trait ChildPathEx {
     fn files(&self) -> Vec<String>;
+    fn assert_is_json<T: DeserializeOwned>(&self) -> T;
 }
 
 impl ChildPathEx for ChildPath {
@@ -48,5 +53,11 @@ impl ChildPathEx for ChildPath {
             .map(|entry| entry.unwrap().file_name().to_string_lossy().into())
             .sorted()
             .collect()
+    }
+
+    fn assert_is_json<T: DeserializeOwned>(&self) -> T {
+        let file = File::open(self.path()).unwrap();
+        let reader = BufReader::new(file);
+        serde_json::from_reader(reader).unwrap()
     }
 }
